feat: update timeline controller to create events per user

+++ src/server/controllers/timeline.controller.js
@@ -2,11 +2,12 @@ const { EventModel } = require('../models/event.model');
 
 // TODO: Perform error handling
 
-const incrementEventCount = async (eventName, data) => {
-  const [event] = await EventModel.find({ name: eventName, data });
+const incrementEventCount = async (userId, eventName, data) => {
+  const [event] = await EventModel.find({ userId, name: eventName, data });
 
   if (!event) {
     await EventModel.create({
+      userId,
       name: eventName,
       count: 1,
       lastUpdated: (new Date()).getTime(),
@@ -14,6 +15,7 @@ const incrementEventCount = async (eventName, data) => {
     });
   } else {
     await EventModel.updateOne({
+      userId,
       name: eventName,
       data,
     }, {
@@ -25,16 +27,19 @@ const incrementEventCount = async (eventName, data) => {
 
 const captureEvent = async (req, res) => {
   const { name, data } = req.body;
+  const { _id } = req.session;
 
   // Force all values to be strings so they can be queried by req.query later on.
   // Clients will be returned the incorrect value types but they should type check anyways.
-  await incrementEventCount(name, Object.fromEntries(Object.entries(data).map(([key, value]) => ([key, value.toString()]))));
+  await incrementEventCount(_id, name, Object.fromEntries(Object.entries(data).map(([key, value]) => ([key, value.toString()]))));
 
   res.status(200).json({ success: true, data: req.body });
 };
 
 const fetchEvents = async (req, res) => {
-  const data = await EventModel.find({}, { _id: 0, __v: 0 });
+  const { _id } = req.session;
+
+  const data = await EventModel.find({ userId: _id }, { _id: 0, __v: 0 });
 
   res.status(200).json({
     success: true,
@@ -46,10 +51,11 @@ const fetchEvents = async (req, res) => {
 
 const fetchEvent = async (req, res) => {
   const { name } = req.params;
+  const { _id } = req.session;
 
   const findQuery = Object.keys(req.query).length
-    ? { name, data: req.query }
-    : { name };
+    ? { userId: _id, name, data: req.query }
+    : { userId: _id, name };
 
   const event = await EventModel.find(findQuery, { _id: 0, __v: 0 });
 
@@ -61,8 +67,9 @@ const fetchEvent = async (req, res) => {
 
 const deleteEvent = async (req, res) => {
   const { name, data } = req.body;
+  const { _id } = req.session;
 
-  await EventModel.deleteOne({ name, data: Object.fromEntries(Object.entries(data).map(([key, value]) => ([key, value.toString()]))) });
+  await EventModel.deleteOne({ userId: _id, name, data: Object.fromEntries(Object.entries(data).map(([key, value]) => ([key, value.toString()]))) });
 
   res.status(200).json({
     success: true,
